screensaver
By default, when the program starts, there are no “control” points and the program is paused. To add points, make a few clicks with the left mouse button anywhere in the program window. The curve will be drawn when there are more than two points on the screen. Press the <P> key to enable curve movement. For help press <F1>.